Bugfix for fuse qconfig comparison
Summary: In some cases the changes to qconfig on module would cause the
fusions to fail. This bugfix solves that problem by adding a
qconfig_function_comparison that compares the functions within the
qconfig rather than the modules the qconfigs are on. The comparison
looks at the partial object within QConfig.activation/weight.p and
compares args, keywords and func. This is necessary to do mannually
because partial doesn't have __eq__ implemented and so == reverts to is.

Test Plan: python test/test_quantization.py
TestFuseFx.test_problematic_fuse_example

Reviewers: supriyar

Subscribers:

Tasks:

Tags:

ghstack-source-id: 5f15e6f7947cba60875e155b3ac56908365f143a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/63384