[Please Ignore] Stacked on #63411
Implement subclass priority for __torch_dispatch__

In order to get this behavior, you have to use append_overloaded,
which I forgot to use in the previous implementation.  I exposed
an internal helper function which is more appropriate for dispatch
to Python where we know that an argument is definitely a Tensor (and
this test no longer needs to be done).

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: e532c53eb5e8e360054edb0c38a69ac23967570c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/63495