Sparse CSR CUDA: add `addmv_out`
This PR adds `addmv_out_sparse_csr_cuda`. The operation is used to
compute matrix-vector multiplication. Since structured_delegate is used
we only need to implement the out variant, the in-place and normal
variants are autogenerated.
Working on this PR revealed that float16 (and probably bfloat16) inputs
do not work correctly in cusparse, therefore for this case `addmm` is
used with squeezes and unsqueezes.

ghstack-source-id: 1b0d6d3b9979b9d2714cbb50a9eb3a95afdedd08
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61407